+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Links in the changelog modal now open in a new tab.
+
 ## [0.5.0] - 2026-06-04
 
 ### Added